Finally fix HIMEM.SYS memory scan.
The last remaining problem was that address flattening was truncating
addresses above 0xFFFF.

+++ include/vomit.h
@@ -57,7 +57,7 @@ extern VomitOptions options;
 
 #define ASSERT_VALID_SEGMENT_INDEX(segmentIndex) VM_ASSERT(static_cast<int>(segmentIndex) >= 0 && static_cast<int>(segmentIndex) <= 5)
 
-inline DWORD vomit_toFlatAddress(WORD segment, WORD offset)
+inline DWORD vomit_toFlatAddress(WORD segment, DWORD offset)
 {
     return (segment << 4) + offset;
 }
